Translation revision
Corrected wrong translation of "Block Text Actions" from "阻止文本操作" to "块文本操作". In Chinese "阻止" is a verb meaning "to stop something", while in this context "block" should have its noun sense translated.

Changed confusing translation of "Gallery" from "画廊" to "图库". This is some what tricky and hard to decide what exact translation to use because the some ROM makers in China have been changing the name of Gallery from "图库"(Gallery) to "相册"(Album) or "照片"(Pictures) and so on, while some ROM makers stick to the name "图库".